GUI & Renderer Updates
Updated GUI to feature a gui_image class for images not fully featured also planning on merging this with the gui_rectangle class

Added a fallback screen to the renderer as before the game crashed when a uninitialized screen was referenced but now it defaults there.

Also removed a bunch of bloat images i left from a earlier project that were just laying around.